docmatcher: store compiled regex
